+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Added
 ~~~~~
 
 - Per-test hooks for modification of data generation strategies. `#492`_
+- Support for ``x-example`` vendor extension in Open API 2.0. `#504`_
 
 `1.1.2`_ - 2020-04-14
 ---------------------
@@ -907,6 +908,7 @@ Fixed
 .. _0.3.0: https://github.com/kiwicom/schemathesis/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/kiwicom/schemathesis/compare/v0.1.0...v0.2.0
 
+.. _#504: https://github.com/kiwicom/schemathesis/issues/504
 .. _#497: https://github.com/kiwicom/schemathesis/issues/497
 .. _#496: https://github.com/kiwicom/schemathesis/issues/496
 .. _#492: https://github.com/kiwicom/schemathesis/issues/492